requirements.txt
pillow==12.0.0
requests==2.32.5
python-dotenv==1.2.1
praw==7.8.1
numpy>=1.26.0,<2.0.0